# [manifold] make partitioner faster. 