--- a/bench/send_bench_one.exs
+++ b/bench/send_bench_one.exs
@@ -1,11 +1,9 @@
 defmodule SendBenchOne do
   use Benchfella
 
-  alias Manifold.Utils
-
   setup_all do
-    pids = [spawn_link &loop/0]
-    {:ok, pids}
+    pid = spawn_link &loop/0
+    {:ok, pid}
   end
 
   defp loop() do
@@ -15,20 +13,19 @@ defmodule SendBenchOne do
   end
 
   bench "send enum each" do
-    bench_context |> Enum.each(&send(&1, :hi))
+    [bench_context] |> Enum.each(&send(&1, :hi))
   end
 
   bench "send list comp" do
-    for pid <- bench_context, do: send(pid, :hi)
+    for pid <- [bench_context], do: send(pid, :hi)
   end
 
   bench "send one" do
-    [pid] = bench_context
-    send(pid, :hi)
+    send(bench_context, :hi)
   end
 
   bench "send fast reducer" do
-    send_r(bench_context, :hi)
+    send_r([bench_context], :hi)
   end
 
   defp send_r([], _msg), do: :ok