--- a/lib/manifold.ex
+++ b/lib/manifold.ex
@@ -33,5 +33,5 @@ defmodule Manifold do
     for {node, pids} <- grouped_by, node != nil, do: Partitioner.send({Partitioner, node}, pids, message)
   end
   def send(pid, message) when is_pid(pid), do: Partitioner.send({Partitioner, node(pid)}, [pid], message)
-  def send(nil, message), do: :ok
+  def send(nil, _message), do: :ok
 end