--- a/lib/manifold/utils.ex
+++ b/lib/manifold/utils.ex
@@ -9,10 +9,25 @@ defmodule Manifold.Utils do
   def group_by(pids, key_fun), do: group_by(pids, key_fun, Map.new)
 
   @spec group_by([pid], key_fun, groups) :: groups
-  def group_by([pid|pids], key_fun, groups) do
+  def group_by([pid | pids], key_fun, groups) do
     key = key_fun.(pid)
     group = Map.get(groups, key) || []
-    group_by(pids, key_fun, Map.put(groups, key, [pid|group]))
+    group_by(pids, key_fun, Map.put(groups, key, [pid | group]))
   end
   def group_by([], _key_fun, groups), do: groups
+
+  @doc """
+  Partitions a bunch of pids into a tuple, of lists of pids grouped by by the result of :erlang.pash2/2
+  """
+  @spec partition_pids([pid], integer) :: tuple
+  def partition_pids(pids, partitions) do
+    do_partition_pids(pids, partitions, Tuple.duplicate([], partitions))
+  end
+
+  defp do_partition_pids([pid | pids], partitions, pids_by_partition) do
+    partition = :erlang.phash2(pid, partitions)
+    pids_in_partition = elem(pids_by_partition, partition)
+    do_partition_pids(pids, partitions, put_elem(pids_by_partition, partition, [pid | pids_in_partition]))
+  end
+  defp do_partition_pids([], _partitions, pids_by_partition), do: pids_by_partition
 end