--- a/lib/manifold/utils.ex
+++ b/lib/manifold/utils.ex
@@ -9,12 +9,12 @@ defmodule Manifold.Utils do
   def group_by(pids, key_fun), do: group_by(pids, key_fun, %{})
 
   @spec group_by([pid], key_fun, groups) :: groups
-  def group_by([pid | pids], key_fun, groups) do
+  defp group_by([pid | pids], key_fun, groups) do
     key = key_fun.(pid)
     group = Map.get(groups, key, [])
     group_by(pids, key_fun, Map.put(groups, key, [pid | group]))
   end
-  def group_by([], _key_fun, groups), do: groups
+  defp group_by([], _key_fun, groups), do: groups
 
   @doc """
   Partitions a bunch of pids into a tuple, of lists of pids grouped by by the result of :erlang.pash2/2
@@ -31,6 +31,9 @@ defmodule Manifold.Utils do
   end
   defp do_partition_pids([], _partitions, pids_by_partition), do: pids_by_partition
 
+  @"""
+  Computes the partition for a given pid using :erlang.phash2/2
+  """
   @spec partition_for(pid, integer) :: integer
   def partition_for(pid, partitions) do
     :erlang.phash2(pid, partitions)