--- a/test/manifold_test.exs
+++ b/test/manifold_test.exs
@@ -6,7 +6,7 @@ defmodule ManifoldTest do
     me = self()
     message = :hello
     pids = for _ <- 0..10000 do
-      spawn fn ->
+      spawn_link fn ->
         receive do
           message -> send(me, {self(), message})
         end
@@ -16,5 +16,17 @@ defmodule ManifoldTest do
     for pid <- pids do
       assert_receive {^pid, ^message}
     end
+  end
+
+  test "send one" do
+    me = self()
+    message = :hello
+    pid = spawn_link fn ->
+      receive do
+        message -> send(me, message)
+      end
+    end
+    Manifold.send([pid], message)
+    assert_receive ^message
   end
 end