--- a/test/manifold_test.exs
+++ b/test/manifold_test.exs
@@ -38,7 +38,7 @@ defmodule ManifoldTest do
         message -> send(me, message)
       end
     end
-    Manifold.send([pid], message)
+    Manifold.send(pid, message)
     assert_receive ^message
   end