@jhgg 2018-02-04
this is clearly faster, i'm wanting to merge this then work on sharding partitions in #4. wdyt?